backend_fastapi/requirements.txt
fastapi[standard]
sqlalchemy
datetime
pyyaml
uvicorn
psycopg2-binary
pg8000
cloud-sql-python-connector[pg8000]
python-dotenv
a2wsgi
sqlmodel==0.0.23
